+++ packages/dart_edge_core/lib/dart_edge_core.dart
@@ -9,6 +9,7 @@ export 'src/client/dart_edge_client_transport.dart';
 export 'src/client/dart_edge_http_client.dart';
 export 'src/context.dart';
 export 'src/http.dart';
+export 'src/isolate_worker.dart';
 export 'src/router.dart';
 export 'src/sql.dart';
 export 'src/web_socket.dart';